+++ src/mt19937-64/Makefile
@@ -164,13 +164,13 @@ am__define_uniq_tagged_files = \
 ETAGS = etags
 CTAGS = ctags
 DISTFILES = $(DIST_COMMON) $(DIST_SOURCES) $(TEXINFOS) $(EXTRA_DIST)
-ACLOCAL = ${SHELL} /home/dorun/research/metasmc-strong/rbedge/missing aclocal-1.13
+ACLOCAL = ${SHELL} /wphsieh/home/dorun/research/metasmc-strong/rbedge/missing aclocal-1.13
 AMTAR = $${TAR-tar}
 AM_DEFAULT_VERBOSITY = 1
 AR = ar
-AUTOCONF = ${SHELL} /home/dorun/research/metasmc-strong/rbedge/missing autoconf
-AUTOHEADER = ${SHELL} /home/dorun/research/metasmc-strong/rbedge/missing autoheader
-AUTOMAKE = ${SHELL} /home/dorun/research/metasmc-strong/rbedge/missing automake-1.13
+AUTOCONF = ${SHELL} /wphsieh/home/dorun/research/metasmc-strong/rbedge/missing autoconf
+AUTOHEADER = ${SHELL} /wphsieh/home/dorun/research/metasmc-strong/rbedge/missing autoheader
+AUTOMAKE = ${SHELL} /wphsieh/home/dorun/research/metasmc-strong/rbedge/missing automake-1.13
 AWK = gawk
 CC = gcc
 CCDEPMODE = depmode=gcc3
@@ -186,17 +186,17 @@ DUMPBIN =
 ECHO_C = 
 ECHO_N = -n
 ECHO_T = 
-EGREP = /bin/grep -E
+EGREP = /usr/bin/grep -E
 EXEEXT = 
-FGREP = /bin/grep -F
-GREP = /bin/grep
+FGREP = /usr/bin/grep -F
+GREP = /usr/bin/grep
 INSTALL = /usr/bin/install -c
 INSTALL_DATA = ${INSTALL} -m 644
 INSTALL_PROGRAM = ${INSTALL}
 INSTALL_SCRIPT = ${INSTALL}
 INSTALL_STRIP_PROGRAM = $(install_sh) -c -s
-LAPACK_CFLAGS = 
-LAPACK_LIBS = -llapack -lblas
+LAPACK_CFLAGS = -I/home/dorun/pkg_miffy/LAPACK/include  
+LAPACK_LIBS = -L/home/dorun/pkg_miffy/LAPACK/lib -llapack -lblas  
 LD = /usr/bin/ld -m elf_x86_64
 LDFLAGS = 
 LIBOBJS = 
@@ -205,9 +205,9 @@ LIBTOOL = $(SHELL) $(top_builddir)/libtool
 LIPO = 
 LN_S = ln -s
 LTLIBOBJS = 
-MAKEINFO = ${SHELL} /home/dorun/research/metasmc-strong/rbedge/missing makeinfo
+MAKEINFO = ${SHELL} /wphsieh/home/dorun/research/metasmc-strong/rbedge/missing makeinfo
 MANIFEST_TOOL = :
-MKDIR_P = /bin/mkdir -p
+MKDIR_P = /usr/bin/mkdir -p
 NM = /usr/bin/nm -B
 NMEDIT = 
 OBJDUMP = objdump
@@ -224,18 +224,18 @@ PACKAGE_VERSION = 1.0
 PATH_SEPARATOR = :
 PKG_CONFIG = /usr/bin/pkg-config
 PKG_CONFIG_LIBDIR = 
-PKG_CONFIG_PATH = /home/dorun/.pkgconfig:/usr/lib/pkgconfig:/usr/share/pkgconfig
+PKG_CONFIG_PATH = /home/dorun/pkgconfig
 POW_LIB = 
 RANLIB = ranlib
-SED = /bin/sed
+SED = /usr/bin/sed
 SET_MAKE = 
-SHELL = /bin/bash
+SHELL = /bin/sh
 STRIP = strip
 VERSION = 1.0
-abs_builddir = /home/dorun/research/metasmc-strong/rbedge/src/mt19937-64
-abs_srcdir = /home/dorun/research/metasmc-strong/rbedge/src/mt19937-64
-abs_top_builddir = /home/dorun/research/metasmc-strong/rbedge
-abs_top_srcdir = /home/dorun/research/metasmc-strong/rbedge
+abs_builddir = /wphsieh/home/dorun/research/metasmc-strong/rbedge/src/mt19937-64
+abs_srcdir = /wphsieh/home/dorun/research/metasmc-strong/rbedge/src/mt19937-64
+abs_top_builddir = /wphsieh/home/dorun/research/metasmc-strong/rbedge
+abs_top_srcdir = /wphsieh/home/dorun/research/metasmc-strong/rbedge
 ac_ct_AR = ar
 ac_ct_CC = gcc
 ac_ct_DUMPBIN = 
@@ -264,7 +264,7 @@ host_vendor = unknown
 htmldir = ${docdir}
 includedir = ${prefix}/include
 infodir = ${datarootdir}/info
-install_sh = ${SHELL} /home/dorun/research/metasmc-strong/rbedge/install-sh
+install_sh = ${SHELL} /wphsieh/home/dorun/research/metasmc-strong/rbedge/install-sh
 libdir = ${exec_prefix}/lib
 libexecdir = ${exec_prefix}/libexec
 localedir = ${datarootdir}/locale
@@ -273,7 +273,7 @@ mandir = ${datarootdir}/man
 mkdir_p = $(MKDIR_P)
 oldincludedir = /usr/include
 pdfdir = ${docdir}
-prefix = /home/dorun/research/metasmc-strong/build
+prefix = /wphsieh/home/dorun/research/metasmc-strong/build
 program_transform_name = s,x,x,
 psdir = ${docdir}
 sbindir = ${exec_prefix}/sbin
