+++ src/smc.c
@@ -11,6 +11,7 @@
 #include "smc.h"
 #include "mutation.h"
 #include "bit.h"
+#include "eindex.h"
 
 double ulast;	// Last uniform random variable generated by abs_time
 size_t nodesize[] = {sizeof(struct coal_node), sizeof(struct migr_node), sizeof(struct xover_node), sizeof(struct sam_node) + sizeof(struct list_head), sizeof(struct dummy_node), sizeof(struct dummy_node)};
@@ -234,569 +235,6 @@ static inline struct edge *edge_set_remove(struct edge_set *set, int i)
 	return e;
 }
 
-void tsindex_rebuild(struct tsindex *tr)
-{
-	double *weights;
-	int i;
-
-	weights = malloc(sizeof(double) * tr->maxnodes);
-	for(i = 0; i < tr->maxnodes; i++){
-		struct edge *e;
-
-		e = tr->edges[i + 1];
-		if(e)
-			weights[i] = e->top->t - e->bot->t;
-		else
-			weights[i] = 0;
-	}
-	__bit_build(tr->index, tr->maxnodes, weights);
-	free(weights);
-	tsindex_clearflag(tr, TSINDEX_DIRTY);
-	tsindex_clearflag(tr, TSINDEX_REBUILD);
-}
-
-struct tsindex *tsindex_alloc(int maxedges)
-{
-	struct tsindex *tr;
-
-	tr = malloc(sizeof(struct tsindex));
-	tr->flags = 0;
-	tr->index = bit_alloc(maxedges);
-	tr->maxedges = maxedges;
-	tr->maxnodes = tr->nedges = 0;
-	tr->edges = malloc(sizeof(struct edge *) * maxedges);
-	memset(tr->edges, 0, sizeof(struct edge *) * maxedges);
-	list_init(&tr->free_list);
-
-	return tr;
-}
-
-void tsindex_reset(struct tsindex *tr)
-{
-	struct list_head *l;
-	struct list *queue;
-
-	tr->flags = 0;
-	memset(tr->edges, 0, sizeof(struct edge *) * tr->maxedges);
-	tr->nedges = tr->maxnodes = 0;
-	bit_clear(tr->index);
-	memset(tr->index->freq, 0, sizeof(double) * tr->index->n);
-
-	/* Clear free list. */
-	queue = &tr->free_list;
-	l = queue->front;
-	while(l){
-		struct list_head *tmp;
-
-		tmp = l->next;
-		__list_remove(queue, l);
-		free(l);
-		l = tmp;
-	}
-}
-
-struct edge *tsindex_search(struct tsindex *tr, double g, double *cum)
-{
-	int eid;
-
-//	if(tsindex_dirty(tr)){
-	if(tsindex_isrebuild(tr)){
-		/* Search operation is disabled when dirty bit is set. */
-		return NULL;
-
-	}else{
-		eid = bit_getindex(tr->index, g);
-		*cum = bit_cumfreq(tr->index, eid - 1);
-
-		return tr->edges[eid];
-	}
-}
-
-double tsindex_size(struct tsindex *tr)
-{
-	if(tsindex_isrebuild(tr))
-		return -1;
-	else
-		return bit_total(tr->index);
-}
-
-void tsindex_add(struct tsindex *tr, struct edge *e)
-{
-	struct list *queue;
-	double diff;
-	int id;
-
-	diff = e->top->t - e->bot->t;
-	queue = &tr->free_list;
-
-	if(queue->front == NULL){
-		/* Allocate a new node in binary indexed tree. */
-		if(tsindex_isrebuild(tr)){
-//			tsindex_setflag(tr, TSINDEX_DIRTY);
-			id = tr->maxnodes + 1;
-
-		}else{
-			id = bit_append(tr->index, diff);
-		}
-		tr->maxnodes++;
-		if(tr->maxnodes >= tr->maxedges){
-			tr->edges = realloc(tr->edges, sizeof(struct edge *) * (tr->maxedges + 1000));
-			memset(tr->edges + tr->maxedges, 0, sizeof(struct edge *) * 1000);
-			tr->maxedges += 1000;
-		}
-
-	}else{
-		struct list_head *l;
-		int *ptr;
-
-		/* Get a existing empty node in binary indexed tree. */
-		l = queue->front;
-		ptr = (int *)GET_OBJ(l);
-		id = *ptr;
-		__list_remove(queue, l);
-		free(l);
-
-		if(!tsindex_isrebuild(tr)){
-//			tsindex_setflag(tr, TSINDEX_DIRTY);
-//
-//		}else{
-			bit_update(tr->index, id, diff);
-		}
-	}
-
-	e->xtid = id;
-	tr->edges[id] = e;
-	tr->nedges++;
-}
-
-void tsindex_update(struct tsindex *tr, struct edge *e, double diff)
-{
-	if(!tsindex_isrebuild(tr))
-//		tsindex_setflag(tr, TSINDEX_DIRTY);
-//	else
-		bit_update(tr->index, e->xtid, diff);
-}
-
-/* Clear a node in binary indexed tree. */
-void tsindex_clear(struct tsindex *tr, struct edge *e)
-{
-	struct list_head *l;
-	int *pidx, id;
-	double diff;
-
-	id = e->xtid;
-	if(!tsindex_isrebuild(tr)){
-//		tsindex_setflag(tr, TSINDEX_DIRTY);
-//
-//	}else{
-		diff = bit_getvalue(tr->index, id);
-		bit_update(tr->index, id, -diff);
-	}
-
-	tr->edges[id] = NULL;
-	e->xtid = 0;
-
-	/* Add freed id to the queue. */
-	l = malloc(sizeof(struct list_head) + sizeof(int));
-	pidx = (int *)GET_OBJ(l);
-	*pidx = id;
-	__list_append(&tr->free_list, l);
-	tr->nedges--;
-}
-
-void tsindex_free(struct tsindex *tr)
-{
-	free(tr->edges);
-	bit_free(tr->index);
-	free(tr);
-}
-
-void tsindex_dump(struct tsindex *tr)
-{
-	struct list_head *l;
-	struct edge *e;
-	int i;
-
-	fprintf(stderr, "tr->flags=%d, tr->nedges=%d, tr->maxnodes=%d, tr->maxedges=%d, G->tr_xoveredges=", tr->flags, tr->nedges, tr->maxnodes, tr->maxedges);
-	for(i = 1; i <= tr->maxnodes; i++){
-		e = tr->edges[i];
-		if(e)
-			fprintf(stderr, "%x(%.10f, %.10f, xtid=%d), ", e, e->bot->t, e->top->t, e->xtid);
-		else
-			fprintf(stderr, "NULL, ");
-	}
-	fprintf(stderr, "\n");
-
-	fprintf(stderr, "free_list=");
-	l = tr->free_list.front;
-	while(l){
-		int *pidx;
-		pidx = (int *)GET_OBJ(l);
-		fprintf(stderr, "%d->", *pidx);
-		l = l->next;
-	}
-	fprintf(stderr, "\n");
-
-	bit_print(tr->index);
-}
-
-/***** LIBAVL allocator of population-wise edge index. *****/
-int eindex_compar(struct edge *a, struct edge *b)
-{
-	double diff, diff2;
-
-	diff = a->top->t - b->top->t;
-	if(diff > 0){
-		return 1;
-
-	}else if(diff < 0){
-		return -1;
-
-	}else{
-		diff2 = a->bot->t - b->bot->t;
-		if(diff2 < 0)
-			return -1;
-		else if(diff2 > 0)
-			return 1;
-		else
-			return a->eid - b->eid;
-	}
-}
-
-void *rbindex_alloc(struct libavl_allocator *allocator, size_t size)
-{
-	return malloc(size);
-}
-
-void rbindex_free(struct libavl_allocator *allocator, void *block)
-{
-	free(block);
-}
-
-struct libavl_allocator rbindex_allocator = {rbindex_alloc, rbindex_free};
-
-static inline void *rbindex_prev(rb_traverser *it)
-{
-	*it = __list_prev(*it);
-	return GET_OBJ(*it);
-}
-
-static inline void *rbindex_cur(rb_traverser it)
-{
-	return GET_OBJ(it);
-}
-
-static inline void *rbindex_next(rb_traverser *it)
-{
-	*it = (*it)->next;
-	return GET_OBJ(*it);
-}
-
-static inline void rbindex_insbefore(rb_traverser it, void *obj)
-{
-	list_insbefore(it, obj);
-}
-
-void rbindex_destroy(struct rbindex *eidx)
-{
-	rb_destroy(eidx->tree, NULL);
-	free(eidx);
-}
-
-void eindex_destroy(struct genealogy *G, struct rbindex *eidx)
-{
-	struct edge *e;
-
-	e = (struct edge *)GET_OBJ(eidx->lsentinel);
-	free_node(G, e->top);
-	free_node(G, e->bot);
-	free_edge(G, e);
-
-	e = (struct edge *)GET_OBJ(eidx->rsentinel);
-	free_node(G, e->top);
-	free_node(G, e->bot);
-	free_edge(G, e);
-
-	rbindex_destroy(eidx);
-}
-
-static inline void **rbindex_insert(struct rbindex *eidx, void *obj)
-{
-	if(rbindex_isseq(eidx)){
-		list_insbefore(eidx->cur_s, obj);
-		eidx->ls.n++;
-
-	}else{
-		void *next;
-
-		next = rb_isam_find(eidx->tree, obj);
-		if(next){
-			list_insbefore(GET_LIST(next), obj);
-			eidx->ls.n++;
-
-		}else{
-			list_append(&eidx->ls, obj);
-		}
-
-		return rb_probe(eidx->tree, obj);
-	}
-}
-
-static inline void rbindex_delete(struct rbindex *eidx, void *obj)
-{
-	if(rbindex_isseq(eidx)){
-		if(GET_LIST(obj) == eidx->cur_s)
-			rbindex_next(&eidx->cur_s);
-		list_remove(&eidx->ls, obj);
-
-	}else{
-		if(rb_delete(eidx->tree, obj))
-			list_remove(&eidx->ls, obj);
-	}
-}
-
-/* Find the first object after the key. */
-static inline void *rbindex_find(rb_traverser *it, struct rbindex *eidx, void *key)
-{
-	void *obj;
-
-	obj = rb_isam_find(eidx->tree, key);
-	eidx->cur_s = *it = GET_LIST(obj);
-	return obj;
-}
-
-static inline int find_root(int n, int h)
-{
-	int half, quad, r;
-
-	half = 1 << (h - 1);
-	quad = half >> 1;
-	r = n - half + 1;
-	if(r < quad)
-		return quad - 1 + r;
-	else
-		return half - 1;
-}
-
-// Calculate indices of complete binary tree from an ordered array
-void complete_binary_tree(int nnodes, int *tree)
-{
-	int i, h, n, half, quad;
-	int *beg, *end, *heights;
-
-	beg = malloc(sizeof(int) * nnodes);
-	end = malloc(sizeof(int) * nnodes);
-	heights = malloc(sizeof(int) * nnodes);
-
-	h = 0;
-	while((1 << h) <= nnodes) h++;
-
-	n = 1;
-	tree[0] = find_root(nnodes, h);
-	beg[0] = 0;
-	end[0] = nnodes;
-	heights[0] = h;
-	i = 0;
-	while(n < nnodes){
-		int j, left, right, r;
-
-		h = heights[i];
-		half = 1 << (h - 1);
-		quad = half >> 1;
-		r = end[i] - beg[i] - half + 1;
-
-		j = tree[i];
-		// Calculate index of left child.
-		if(n < nnodes){
-			heights[n] = h - 1;
-			left = beg[i] + find_root(j - beg[i], heights[n]);
-			tree[n] = left;
-			beg[n] = beg[i];
-			end[n] = j;
-			n++;
-		}
-		// Calculate index of left child.
-		if(n < nnodes){
-			if(r < quad)
-				heights[n] = h - 2;
-			else
-				heights[n] = h - 1;
-			right = j + 1 + find_root(end[i] - j - 1, heights[n]);
-			tree[n] = right;
-			beg[n] = j + 1;
-			end[n] = end[i];
-			n++;
-		}
-		i++;
-	}
-
-	free(heights);
-	free(end);
-	free(beg);
-}
-
-// Rebuild tree index from sorted list
-void rbindex_rebuild_tree(struct rbindex *eidx)
-{
-	int i, nnodes, h, half, *tree;
-	struct rb_node **nodes;
-	struct list_head *l;
-	void **objs;
-
-	nnodes = eidx->ls.n;
-	rb_destroy(eidx->tree, NULL);
-	objs = malloc(sizeof(void *) * nnodes);
-	l = eidx->ls.front;
-	for(i = 0; i < nnodes; i++){
-		objs[i] = GET_OBJ(l);
-		l = l->next;
-	}
-
-	// Build complete binary tree
-	tree = malloc(sizeof(int) * nnodes);
-	complete_binary_tree(nnodes, tree);
-	h = 0;
-	while(1 << h <= nnodes) h++;
-	half = 1 << (h - 1);
-
-	nodes = malloc(sizeof(struct rb_node *) * nnodes);
-	for(i = nnodes - 1; i >= 0; i--){
-		nodes[i] = malloc(sizeof(struct rb_node));
-		nodes[i]->rb_data = objs[tree[i]];
-
-		// Color nodes
-		if(i < half - 1)
-			nodes[i]->rb_color = RB_BLACK;
-		else
-			nodes[i]->rb_color = RB_RED;
-
-		// Link nodes
-		if(i * 2 + 1 >= nnodes)
-			nodes[i]->rb_link[0] = NULL;
-		else
-			nodes[i]->rb_link[0] = nodes[i * 2 + 1];
-		if(i * 2 + 2 >= nnodes)
-			nodes[i]->rb_link[1] = NULL;
-		else
-			nodes[i]->rb_link[1] = nodes[i * 2 + 2];
-	}
-
-	eidx->tree = rb_create(eidx->compar, NULL, &rbindex_allocator);
-	eidx->tree->rb_root = nodes[0];
-	nodes[0]->rb_color = RB_BLACK;
-
-	free(nodes);
-	free(objs);
-	free(tree);
-}
-
-struct rbindex *rbindex_create(rb_comparison_func *compar, void *param, struct libavl_allocator *allocator)
-{
-	struct rbindex *eidx;
-
-	eidx = malloc(sizeof(struct rbindex));
-	eidx->flags = 0;
-	eidx->compar = compar;
-	eidx->tree = rb_create(compar, param, allocator);
-	list_init(&eidx->ls);
-
-	return eidx;
-}
-
-/* Set up cursor. */
-static inline void rbindex_s_set(struct rbindex *eidx, void *obj)
-{
-	eidx->cur_s = GET_LIST(obj);
-}
-
-static inline rb_traverser rbindex_s_get(void *obj)
-{
-	return GET_LIST(obj);
-}
-
-static inline void rbindex_s_reset(struct rbindex *eidx)
-{
-	eidx->cur_s = eidx->ls.front->next;
-}
-
-/* Turn on sequential mode. */
-void rbindex_seq_on(struct rbindex *eidx)
-{
-	rbindex_setflag(eidx, RBINDEX_SEQUENTIAL);
-}
-
-/* Turn off sequential mode. */
-void rbindex_seq_off(struct rbindex *eidx)
-{
-	rbindex_clearflag(eidx, RBINDEX_SEQUENTIAL);
-	rbindex_rebuild_tree(eidx);
-}
-
-int rbindex_s_first(struct rbindex *eidx, rb_traverser cur)
-{
-	return cur == &eidx->ls.front;
-}
-
-void eindex_s_seek(struct rbindex *eidx, double t)
-{
-	if(rbindex_isseq(eidx)){
-		rb_traverser lfwd, lbwd;
-		struct edge *ebwd, *efwd;
-
-		lbwd = eidx->cur_s;
-		ebwd = rbindex_cur(lbwd);
-		while(ebwd->top->t >= t) ebwd = rbindex_prev(&lbwd);
-		lfwd = lbwd;
-		efwd = rbindex_cur(lfwd);
-		while(efwd->top->t < t) efwd = rbindex_next(&lfwd);
-		eidx->cur_s = lfwd;
-	}
-}
-
-static inline void eindex_insert(struct rbindex *eidx, struct edge *e)
-{
-	if(rbindex_isseq(eidx)){
-		rb_traverser cur;
-		struct edge *efwd;
-
-		eindex_s_seek(eidx, e->top->t);
-		cur = eidx->cur_s;
-		efwd = rbindex_cur(cur);
-		while(efwd->top->t == e->top->t && efwd->bot->t < e->bot->t) efwd = rbindex_next(&cur);
-		while(efwd->top->t == e->top->t && efwd->bot->t == e->bot->t && efwd->eid < e->eid) efwd = rbindex_next(&cur);
-
-		eidx->cur_s = cur;
-	}
-
-	rbindex_insert(eidx, (void *)e);
-}
-
-static inline struct rbindex *eindex_create(struct genealogy *G, int pop)
-{
-	struct rbindex *eidx;
-	struct node *top, *bot;
-	struct edge *e;
-
-	eidx = rbindex_create(eindex_compar, NULL, &rbindex_allocator);
-
-	// Set up sentinel
-	top = alloc_node(G, NODE_DUMMY, pop, 0);
-	bot = alloc_node(G, NODE_DUMMY, pop, 0);
-	e = alloc_edge(G, top, bot);
-	eidx->lsentinel = GET_LIST(e);
-
-	list_append(&eidx->ls, e);
-
-	top = alloc_node(G, NODE_DUMMY, pop, INFINITY);
-	bot = alloc_node(G, NODE_DUMMY, pop, INFINITY);
-	e = alloc_edge(G, top, bot);
-	eidx->rsentinel = GET_LIST(e);
-
-	list_append(&eidx->ls, e);
-
-	return eidx;
-}
-
 void
 print_tree_structure (struct rb_node *node, int level)
 {
@@ -989,7 +427,6 @@ void insert_migr_node(struct genealogy *G, struct edge *e, struct migr_node *nd)
 	nd->in = e;
 	nd->out = e2;
 	add_edge(G, pop, e2);
-//	eindex_insert(G->pops[e2->bot->pop].eidx, e2);
 }
 
 /* Erase dangling lineage until a coalescent node is reached. */
@@ -1009,18 +446,18 @@ void erase_dangling2(struct genealogy *G, struct edge *e)
 	/* Remove dangling edges from red-black trees. */
 	etmp = e;
 	while(etmp->top->type == NODE_MIGR){
-		rbindex_delete(G->pops[etmp->bot->pop].eidx, etmp);
+		eindex_delete(G->pops[etmp->bot->pop].eidx, etmp);
 		etmp = etmp->top->in;
 	}
-	rbindex_delete(G->pops[etmp->bot->pop].eidx, etmp);
+	eindex_delete(G->pops[etmp->bot->pop].eidx, etmp);
 	ntop = (struct coal_node *)etmp->top;
 #ifdef DEBUG
 	fprintf(stderr, "eabove=%x, ebelow=%x, ein=%x\n", ntop->in, ntop->out[1 - etmp->itop], etmp);
 #endif
 	eabove = ntop->in;
-	rbindex_delete(G->pops[eabove->bot->pop].eidx, eabove);
+	eindex_delete(G->pops[eabove->bot->pop].eidx, eabove);
 	ebelow = ntop->out[1 - etmp->itop];
-	rbindex_delete(G->pops[ebelow->bot->pop].eidx, ebelow);
+	eindex_delete(G->pops[ebelow->bot->pop].eidx, ebelow);
 
 	/* Remove dangling edges. */
 	pop = e->bot->pop;
@@ -1137,7 +574,7 @@ rb_traverser choose_tedge(struct genealogy *G, struct population *pop, double t)
 			bot.t = 0;
 			key.top = &top;
 			key.bot = &bot;
-			e = rbindex_find(&tr, pop->eidx, &key);
+			e = eindex_find(&tr, pop->eidx, &key);
 #ifdef DEBUG
 			fprintf(stderr, " Using red-black index: key=(%.6f, %.6f), start from e=(%6f, %.6f)\n", key.bot->t, key.top->t, e->bot->t, e->top->t);
 #endif
@@ -1152,7 +589,7 @@ rb_traverser choose_tedge(struct genealogy *G, struct population *pop, double t)
 				else
 					break;
 			}
-			e = rbindex_next(&tr);
+			e = eindex_next(&tr);
 		}
 #ifdef DEBUG
 		fprintf(stderr, "Choosed: (%x(%.6f, %d), %x(%.6f, %d))\n", e->bot, e->bot->t, e->bot->pop, e->top, e->top->t, e->top->pop);
@@ -1178,7 +615,7 @@ struct event *__absorption(struct genealogy *G, struct edge *e, struct edge *f, 
 //	nabs++;
 	fprintf(stderr, "Entering function %s: e=%x, f=%x, pop=%d, t=%.6f\n", __func__, e, f, pop, t);
 #endif
-//	rbindex_delete(G->pops[e->bot->pop].eidx, e);
+//	eindex_delete(G->pops[e->bot->pop].eidx, e);
 	nd = (struct coal_node *)alloc_node(G, NODE_COAL, pop, t);
 
 	if(t > G->localMRCA->t){
@@ -1265,10 +702,10 @@ struct event *absorption(struct genealogy *G, struct edge *f, int pop, double t)
 		e = (struct edge *)GET_OBJ(cur);
 //	}
 
-	rbindex_s_set(G->pops[e->bot->pop].eidx, e);
+	eindex_s_set(G->pops[e->bot->pop].eidx, e);
 	
 	// Remove e from the index
-	rbindex_delete(G->pops[e->bot->pop].eidx, e);
+	eindex_delete(G->pops[e->bot->pop].eidx, e);
 
 	ev = __absorption(G, e, f, pop, t);
 	e_new = AS_COAL_NODE(e->bot)->out[0];	// Get the new edge allocated by insert_coal_node (Old lineage below coalescent node).
@@ -1282,11 +719,6 @@ struct event *absorption(struct genealogy *G, struct edge *f, int pop, double t)
 	/* Reinsert e into the edge list. */
 //	eindex_s_seek(G->pops[e->bot->pop].eidx, t);
 	eindex_insert(G->pops[e->bot->pop].eidx, e);
-
-//	cur = G->pops[e->bot->pop].eidx->cur_s;
-//	efwd = rbindex_cur(cur);
-//	while(efwd->top->t >= e->top->t && efwd->bot->t < e->bot->t && efwd->eid < e->eid) efwd = rbindex_next(&cur);
-//	rbindex_insbefore(cur, e);
 
 /*	if(t > tmrca_old){
 		free_node(G, e->top);
@@ -1678,7 +1110,7 @@ void reset_populations(struct genealogy *G)
 
 		for(; i < cfg->npop + cfg->nsplt; i++)
 			G->pops[pop].mrate[i] = 0;
-		rbindex_s_reset(G->pops[pop].eidx);
+		eindex_s_reset(G->pops[pop].eidx);
 	}
 
 	/* Reset populations created by splt event. */
@@ -1689,7 +1121,7 @@ void reset_populations(struct genealogy *G)
 		G->pops[pop].tlast = 0;
 		G->pops[pop].enabled = 0;
 		memset(G->pops[pop].mrate, 0, sizeof(double) * (cfg->npop + cfg->nsplt));
-		rbindex_s_reset(G->pops[pop].eidx);
+		eindex_s_reset(G->pops[pop].eidx);
 	}
 }
 
@@ -1890,7 +1322,7 @@ double recombination(struct genealogy *G)
 	fprintf(stderr, "Next event: evl=%x, ev=%x, type=%d, t=%6f\n", evl, ev, ev->type, ev->t);
 #endif
 
-	rbindex_delete(G->pops[e->bot->pop].eidx, e);
+	eindex_delete(G->pops[e->bot->pop].eidx, e);
 
 	/* Generate recombination node */
 	nxover = (struct xover_node *)alloc_node(G, NODE_XOVER, pop, t);
@@ -1955,7 +1387,7 @@ double recombination(struct genealogy *G)
 
 				}else{*/
 					cur = choose_tedge(G, &G->pops[pop], t);
-					e2 = rbindex_cur(cur);
+					e2 = eindex_cur(cur);
 //				}
 
 
@@ -1963,9 +1395,9 @@ double recombination(struct genealogy *G)
 				fprintf(stderr, "%d: Absorb floating lineage %x to %x, e=%x\n", __LINE__, ef, e2, e);
 #endif
 
-				rbindex_delete(G->pops[nxover->pop].eidx, nxover->in);
+				eindex_delete(G->pops[nxover->pop].eidx, nxover->in);
 
-				rbindex_delete(G->pops[e2->bot->pop].eidx, e2);
+				eindex_delete(G->pops[e2->bot->pop].eidx, e2);
 
 				evnew = __absorption(G, e2, ef, pop, t);
 
@@ -1974,14 +1406,14 @@ double recombination(struct genealogy *G)
 				eindex_insert(G->pops[pop].eidx, AS_COAL_NODE(e2->bot)->out[0]);
 
 				list_insbefore(evl, evnew);
-				rbindex_delete(G->pops[nxover->pop].eidx, nxover->in_new);
+				eindex_delete(G->pops[nxover->pop].eidx, nxover->in_new);
 
 				if(e2 == e){	// Floating lineage is absorbed to the same lineage (loop in ARG)
 					struct edge *e_old;
 
-					rbindex_delete(G->pops[pop].eidx, e);
-					rbindex_delete(G->pops[pop].eidx, AS_COAL_NODE(e2->bot)->out[0]);
-					rbindex_delete(G->pops[pop].eidx, AS_COAL_NODE(e2->bot)->out[1]);
+					eindex_delete(G->pops[pop].eidx, e);
+					eindex_delete(G->pops[pop].eidx, AS_COAL_NODE(e2->bot)->out[0]);
+					eindex_delete(G->pops[pop].eidx, AS_COAL_NODE(e2->bot)->out[1]);
 
 					nxover->in = AS_COAL_NODE(e2->bot)->out[0];
 					e_old = nxover->in;
@@ -2003,7 +1435,7 @@ double recombination(struct genealogy *G)
 					struct edge *e_new, *e_below;
 
 					e_new = nxover->in_new;
-					rbindex_delete(G->pops[e_new->bot->pop].eidx, e_new);
+					eindex_delete(G->pops[e_new->bot->pop].eidx, e_new);
 					e_below = nxover->out;
 					tsindex_update(G->tr_xover, e_new, e_below->top->t - e_below->bot->t);
 					e_new->bot = e_below->bot;
@@ -2083,8 +1515,8 @@ double recombination(struct genealogy *G)
 
 				troot_old = G->root->t;
 
-				rbindex_delete(G->pops[nxover->pop].eidx, nxover->in);
-				rbindex_delete(G->pops[nxover->pop].eidx, nxover->in_new);
+				eindex_delete(G->pops[nxover->pop].eidx, nxover->in);
+				eindex_delete(G->pops[nxover->pop].eidx, nxover->in_new);
 
 #ifdef DEBUG
 				if(ev->type == EVENT_DUMY){
@@ -2104,7 +1536,7 @@ double recombination(struct genealogy *G)
 					edum = ndum->in;
 					ndum = edum->top;
 				}
-				rbindex_delete(G->pops[edum->bot->pop].eidx, edum);
+				eindex_delete(G->pops[edum->bot->pop].eidx, edum);
 
 				// Remove remaining edges in dummy edge list
 				if(ev->type == EVENT_DXVR){
@@ -2115,7 +1547,7 @@ double recombination(struct genealogy *G)
 					/* Remove dangling edges above dummy recombination event from red-black tree. */
 					erm = ndum->in;
 					while(erm){
-						rbindex_delete(G->pops[erm->bot->pop].eidx, erm);
+						eindex_delete(G->pops[erm->bot->pop].eidx, erm);
 						erm = erm->top->in;
 					}
 
@@ -2167,8 +1599,8 @@ double recombination(struct genealogy *G)
 				{
 					struct edge *e_new, *e_below;
 
-					rbindex_delete(G->pops[nxover->pop].eidx, nxover->in);
-					rbindex_delete(G->pops[nxover->pop].eidx, nxover->in_new);
+					eindex_delete(G->pops[nxover->pop].eidx, nxover->in);
+					eindex_delete(G->pops[nxover->pop].eidx, nxover->in_new);
 					erase_dangling2(G, nxover->in);
 
 					e_new = nxover->in_new;
@@ -2309,7 +1741,7 @@ double merge_floating(struct genealogy *G, struct edge_set *F)
 
 				t += minu;
 				evnew = absorption(G, e, upop, t);
-				rbindex_s_set(G->pops[upop].eidx, e);
+				eindex_s_set(G->pops[upop].eidx, e);
 				eindex_s_seek(G->pops[upop].eidx, t);
 				sumnF--;
 
@@ -2319,7 +1751,7 @@ double merge_floating(struct genealogy *G, struct edge_set *F)
 				sublike = log(2) - totalprob * minv;
 				nd = coalescent(G, F, vpop, t);
 				evnew = (struct event *)nd->ev;
-				rbindex_s_set(G->pops[vpop].eidx, nd->out[0]);
+				eindex_s_set(G->pops[vpop].eidx, nd->out[0]);
 				eindex_s_seek(G->pops[vpop].eidx, t);
 				sumnF--;
 
@@ -2355,7 +1787,7 @@ finish_selection:
 //				fprintf(stdout, "nmigr=%d, dpop=%d, c=%d, ev->dpop=%d, ev->spop=%d\n", nmigr, dpop, c, ev->dpop, ev->spop);
 
 				sublike = log(G->pops[zpop].mrate[zpop]) - totalprob * minz;
-				rbindex_s_set(G->pops[zpop].eidx, nd->out);
+				eindex_s_set(G->pops[zpop].eidx, nd->out);
 				eindex_s_seek(G->pops[zpop].eidx, t);
 			}
 
@@ -2432,7 +1864,7 @@ finish_selection:
 					edum = ndum->in;
 					ndum = edum->top;
 				}
-				rbindex_delete(G->pops[edum->bot->pop].eidx, edum);
+				eindex_delete(G->pops[edum->bot->pop].eidx, edum);
 
 				// Remove remaining edges in dummy edge list
 				if(ev->type == EVENT_DXVR){
@@ -2443,7 +1875,7 @@ finish_selection:
 					/* Remove dangling edges above dummy recombination event from red-black tree. */
 					erm = ndum->in;
 					while(erm){
-						rbindex_delete(G->pops[erm->bot->pop].eidx, erm);
+						eindex_delete(G->pops[erm->bot->pop].eidx, erm);
 						erm = erm->top->in;
 					}
 
@@ -3016,7 +2448,7 @@ int simulate(struct genealogy *G, struct profile *prof)
 		tsindex_setflag(G->tr_xover, TSINDEX_REBUILD);
 		// Enter sequential mode.
 		for(i = 0; i < cfg->npop_all; i++)
-			rbindex_seq_on(G->pops[i].eidx);
+			eindex_seq_on(G->pops[i].eidx);
 
 		sublike = merge_floating(G, F);
 
@@ -3030,7 +2462,7 @@ int simulate(struct genealogy *G, struct profile *prof)
 
 		// Leave sequential mode and rebuild red-black tree index
 		for(i = 0; i < cfg->npop_all; i++)
-			rbindex_seq_off(G->pops[i].eidx);
+			eindex_seq_off(G->pops[i].eidx);
 		tsindex_rebuild(G->tr_xover);
 
 #ifdef DEBUG
